+++ src/threads.js
@@ -65,7 +65,7 @@ StagePickerMorph, CustomBlockDefinition*/
 
 /*jshint esversion: 11, bitwise: false, evil: true*/
 
-modules.threads = '2023-February-15';
+modules.threads = '2023-February-16';
 
 var ThreadManager;
 var Process;
@@ -873,67 +873,31 @@ Process.prototype.reportApplyExtension = function (prim, args) {
 // Process: Special Forms Blocks Primitives
 
 Process.prototype.reportVariadicOr = function (block) {
-    var inputs = this.context.inputs,
-        tests = block.inputs()[0],
-        inline = tests instanceof MultiArgMorph,
-        outer = this.context.outerContext,
-        acc = this.context.accumulator;
-    if (inputs.length < 1) {
-        if (inline) {
-            acc = this.context.accumulator = {
-                slots: tests.inputs(),
-                len: tests.inputs().length,
-                pc: 1
-            };
-            this.pushContext(acc.slots[0], outer);
-        } else { // tests is an ArgLabelMorph
-            this.pushContext(tests.argMorph(), outer);
-        }
-    } else if (inputs.length === 1) {
-        if (acc) { // inline - acc has been initialized
-            if (inputs[0] === true) {
-                if (this.flashContext()) {return; }
-                this.returnValueToParentContext(true);
-                this.popContext();
-            } else if (acc.pc === acc.len) {
-                if (this.flashContext()) {return; }
-                this.returnValueToParentContext(inputs[0]);
-                this.popContext();
-            } else {
-                if (inline) {
-                    this.pushContext(acc.slots[acc.pc], outer);
-                } else {
-                    this.pushContext();
-                    this.evaluate(acc.slots[acc.pc], new List());
-                }
-            }
-        } else { // "with input list" variant
-            this.context.accumulator = {
-                slots: inputs[0].itemsArray(),
-                len: inputs[0].length(),
-                pc: 1
-            };
-            inputs.pop();
-            acc = this.context.accumulator;
-            this.pushContext();
-            this.evaluate(acc.slots[0], new List());
-        }
-    } else {
-        if (this.flashContext()) {return; }
-        inputs.push(this.hyper(
-            this.reportBasicOr,
-            inputs.pop(),
-            inputs.pop()
-        ));
-        acc.pc += 1;
-    }
-};
-
-Process.prototype.reportBasicOr = function (a, b) {
-    return a || b;
+    this.reportAssociativeBool(
+        block,
+        this.reportBasicOr,
+        value => value === true,
+        true
+    );
 };
 
 Process.prototype.reportVariadicAnd = function (block) {
+    this.reportAssociativeBool(
+        block,
+        this.reportBasicAnd,
+        value => value !== true,
+        false
+    );
+};
+
+Process.prototype.reportAssociativeBool = function (
+    // private - evaluate special form variadic associative Boolean operations
+    // such as AND, OR
+    block,
+    baseOp, // dyadic base operation (AND, OR)
+    shortCircuit, // monadic predicate
+    shortCut // return value if shortCircuit evaluates to TRUE
+) {
     var inputs = this.context.inputs,
         tests = block.inputs()[0],
         inline = tests instanceof MultiArgMorph,
@@ -952,9 +916,9 @@ Process.prototype.reportVariadicAnd = function (block) {
         }
     } else if (inputs.length === 1) {
         if (acc) { // inline - acc has been initialized
-            if (!inputs[0]) {
+            if (shortCircuit(inputs[0])) {
                 if (this.flashContext()) {return; }
-                this.returnValueToParentContext(false);
+                this.returnValueToParentContext(shortCut);
                 this.popContext();
             } else if (acc.pc === acc.len) {
                 if (this.flashContext()) {return; }
@@ -982,12 +946,15 @@ Process.prototype.reportVariadicAnd = function (block) {
     } else {
         if (this.flashContext()) {return; }
         inputs.push(this.hyper(
-            this.reportBasicAnd,
+            baseOp,
             inputs.pop(),
             inputs.pop()
         ));
         acc.pc += 1;
     }
+};
+Process.prototype.reportBasicOr = function (a, b) {
+    return a || b;
 };
 
 Process.prototype.reportBasicAnd = function (a, b) {