+++ HISTORY.md
@@ -15,6 +15,7 @@
 * adjusted Bulgarian translation for Boolean ops
 * blocks: fixed "rename all" for custom block formal parameters
 * blocks: fixed "rename" for custom block formal parameters
+* threads: refactored variadic AND, OR into generalized reportAssociativeBool()
 
 ### 2023-02-15
 * pushed dev version to 8.2